+++ src/Layout/AssetGrid.cpp
@@ -14,7 +14,7 @@ AssetGrid::AssetGrid( Vector2& _cursor_position, Vector2 _size )
 : UIElement( _cursor_position, _size )
 {
 	m_asset_grid_size.x  = m_inner_size.x - UI::MARGIN * 2;
-	m_grid_render_target = LoadRenderTexture( m_inner_size.x, m_inner_size.y );
+	m_grid_render_target = LoadRenderTexture( static_cast< int >( m_inner_size.x ), static_cast< int >( m_inner_size.y ) );
 	auto frame_width     = ( ( _size.x - ( UI::MARGIN * ( FRAME_ROW_COUNT + 1 ) ) ) / FRAME_ROW_COUNT );
 	m_frame_size         = Vector2( frame_width, frame_width * FRAME_ASPECT_RATIO );
 }
@@ -55,7 +55,7 @@ void AssetGrid::update(void)
 void AssetGrid::drawAssetGrid( void )
 {
 	BeginTextureMode( m_grid_render_target );
-	ClearBackground( Color( 0.f, 0.f, 0.f, 0.f ) );
+	ClearBackground( Color( 0, 0, 0, 0 ) );
 
 	Vector2 reset_position  = Vector2( 0.f, 0.f );
 	Vector2 cursor_position = reset_position;
@@ -95,8 +95,8 @@ void AssetGrid::drawInner( Vector2 _cursor_position )
 	RaylibRectangle rect;
 	rect.x      = 0.f;
 	rect.y      = 0.f;
-	rect.height = -m_grid_render_target.texture.height;
-	rect.width  = m_grid_render_target.texture.width;
+	rect.height = static_cast< float >( -m_grid_render_target.texture.height );
+	rect.width  = static_cast< float >( m_grid_render_target.texture.width );
 	DrawTextureRec( m_grid_render_target.texture, rect, _cursor_position, Color( 255, 255, 255, 255 ) );
 }
